LayoutTests/fast/canvas/webgl/drawElements-empty-vertex-data-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawElements: no buffer is bound to enabled attribute
PASS: Unable to draw with invalid vertexAttribArray0
